+++ android/expoview/src/main/java/host/exp/exponent/experience/ExperienceActivity.java
@@ -415,6 +415,7 @@ public class ExperienceActivity extends BaseExperienceActivity implements Expone
       mManagedAppSplashScreenViewProvider.updateSplashScreenViewWithManifest(this, manifest);
     }
   }
+  
   public void setLoadingProgressStatusIfEnabled() {
     ExpoUpdatesAppLoader appLoader = mKernel.getAppLoaderForManifestUrl(mManifestUrl);
     if (appLoader != null) {