+++ packages/expo-splash-screen/CHANGELOG.md
@@ -9,6 +9,8 @@
 - Add warning for splash screen if visible for too long ([#12882](https://github.com/expo/expo/pull/12882) by [@ajsmth](https://github.com/ajsmth))
 ### 🐛 Bug fixes
 
+- On iOS, search for a view controller with a RCTRootView rather than always using the keyWindow's rootViewController. ([#13429](https://github.com/expo/expo/pull/13429) by [@esamelson](https://github.com/esamelson))
+
 ### 💡 Others
 
 ## 0.11.0 — 2021-06-16